fortran file to post-process LOCPOT
A modified version of vtotav.f to post-process (to do the planar average along the normal direction of a surface) the LOCPOT, which is generated by the VASP4.x or VASP5.x versions. The output file of this tool is VLINE, which can be used to obtain the vacuum energy level. Then the work function of the calculated surface can be obtained by the Fermi level and the vacuum energy level.